+++ src/app/App.svelte
@@ -180,6 +180,7 @@
   })
 
   router.register("/", Home)
+  router.register("/topics/:topic", Home)
   router.register("/notes", Home)
   router.register("/notes/create", NoteCreate, {
     requireSigner: true,