+++ src/app/shared/NoteContentKind9802.svelte
@@ -18,8 +18,8 @@
 </script>
 
 <div class="flex flex-col gap-2">
-  {#if meta.comment}
-    <NoteContentKind1 note={{content: meta.comment}} {showMedia} {showEntire} />
+  {#if meta.context}
+    <NoteContentKind1 note={{content: meta.context}} {showMedia} {showEntire} />
   {/if}
   <div class="flex flex-col gap-2 overflow-hidden text-ellipsis">
     <div class="border-l-2 border-solid border-neutral-600 pl-4">